TileSystem/CardCoreUpdates
Implemented three types of cards, one for the demoman, one for dark princess, and one for a test spell. The Card script contains references to the placing script, the object associated with the card, has a name and a qweix cost associated with it. I added a rudimentary Qweix count with buttons for adding and removing qweix, to implement cost for spawning.